// src/shared/pet-interfaces.ts
export interface Pet {
    id: number
    shelter_id: number
    name: string
    location: string
    breed: string
    age: number
    gender: string
    size: number
    color: string
    status: string
    behavior: string
    description: string
    medical_history: string
    special_needs: string;
    shelter_user_id: number;
    shelter_name: string;
}